Important bug fix
Looks like there's a bug in angular that makes the forms not validate
properly the getMaxDate function is modified. I removed my date
modification code to make the dates work with the bug. Maybe this code
can be added back in in the future when this is working properly in
angular.